requirements.txt
flask
kafka-python
redis